+++ assets/Changelog.md
@@ -25,6 +25,8 @@ Added support for UE Version 5.5 - ([UE4SS #708](https://github.com/UE4SS-RE/RE-
 
 Added support for UE Version 5.4 - ([UE4SS #503](https://github.com/UE4SS-RE/RE-UE4SS/pull/503)) 
 
+Added support for UE Versions 4.7 through 4.10 - ([UE4SS #1390](https://github.com/UE4SS-RE/RE-UE4SS/pull/1390))
+
 Added basic support for Development/Debug/Test built Unreal Engine games ([UE4SS #607](https://github.com/UE4SS-RE/RE-UE4SS/pull/607)) 
 - To use this functionality, set DebugBuild to true in UE4SS-Settings.ini