+++ assets/Changelog.md
@@ -109,6 +109,8 @@ Fixed the ingame console (Tilde, or F10 key) not working for some games - ([UE4S
 
 Fixed random crashes (use-after-free) caused by the string intern pool storing dangling pointers to caller-owned strings, most commonly Lua GC'd strings from UObject member access ([UE4SS #1271](https://github.com/UE4SS-RE/RE-UE4SS/pull/1271))
 
+Added `OnUnload` lua function, register callbacks to be called when the mods is uninstalled / reloaded
+
 ### Live View 
 Added search filter: `IncludeClassNames`. ([UE4SS #472](https://github.com/UE4SS-RE/RE-UE4SS/pull/472)) - Buckminsterfullerene